not a task feature yet: some authentication improvements
