+++ vite.config.js
@@ -4,11 +4,9 @@ import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import VueDevTools from 'vite-plugin-vue-devtools'
 
-import envPlugin from 'vite-plugin-env'
-
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue(), VueDevTools(), envPlugin()],
+  plugins: [vue(), VueDevTools()],
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url))
